Commit modify functionality.
Commit modify functionality

+++ VUTTRFONT/src/app/component/template/cardsrcnadd/cardsrcnadd.component.ts
@@ -3,6 +3,8 @@ import { MatDialog } from '@angular/material/dialog';
 import { DialognewComponent } from 'src/app/component/template/dialognew/dialognew.component'
 import { Tool } from 'src/app/tool.model';
 import { ToolService } from '../../service/tool.service';
+import { DialogdeleteComponent } from '../dialogdelete/dialogdelete.component';
+import { DialogmodifyComponent } from '../dialogmodify/dialogmodify.component';
 @Component({
   selector: 'app-cardsrcnadd',
   templateUrl: './cardsrcnadd.component.html',
@@ -13,6 +15,7 @@ export class CardsrcnaddComponent implements OnInit {
   tools : Tool[] = [];
   searchType: boolean = false;
   stringSearch: string = "";
+  delete: boolean = false;
 
   constructor(
     public dialog: MatDialog,
@@ -34,13 +37,44 @@ export class CardsrcnaddComponent implements OnInit {
     });
 
     dialogRef.afterClosed().subscribe(result => {
-      console.log('The dialog was closed');
+      console.log(result);
+      this.service.create(result).subscribe(() =>{
+        this.service.showMessage("Tool created with Success!");
+        this.atualizarDados();
+      },
+      err => {
+        this.service.showMessage("It was not possible to create tool.")
+      });
+    });
+  }
+  modifyTool(tool: Tool): void{
+    const dialogRef = this.dialog.open(DialogmodifyComponent, {
+      width: '400px',
+      data: tool,
+    });
+    dialogRef.afterClosed().subscribe(result => {
+      console.log(result);
+      this.service.update(result).subscribe(() =>{
+        this.service.showMessage("Tool modified with Success!");
+        this.atualizarDados();
+      },
+      err => {
+        this.service.showMessage("It was not possible to modify tool.")
+      });
     });
   }
   deleteTool(tool: Tool): void{
-    this.service.delete(tool).subscribe(() =>  {
-      this.service.showMessage("Tool Deleted with success");
-      this.atualizarDados();
+    
+    const dialogRef = this.dialog.open(DialogdeleteComponent, {
+      width: '400px',
+    });
+    dialogRef.afterClosed().subscribe(response => {
+      if(response == true){
+        this.service.delete(tool).subscribe(() =>  {
+          this.service.showMessage("Tool Deleted with success");
+          this.atualizarDados();
+        });
+      };
     });
   }
   //Lógica para buscar ferramenta de acordo com os parâmetros de seleção
@@ -67,5 +101,4 @@ export class CardsrcnaddComponent implements OnInit {
     this.searchType = false;
   }
 }
-
